Changed port IDfrom 32 to 16 bits in RestServerArgsParser.

+++ project/server/src/common/args/RestServerArgsParser.cpp
@@ -40,7 +40,7 @@ RestServerArgsParser::RestServerArgsParser(const std::vector<std::string>& args)
     }
 }
 
-uint32_t RestServerArgsParser::parsePort_(const std::string& portString) const {
+uint16_t RestServerArgsParser::parsePort_(const std::string& portString) const {
     uint32_t portId;
     std::istringstream iStrStrm(portString);
     iStrStrm >> portId;
